SetFieldsOnce really only sets the field once
When added, @fields was cached in a %once block and adding more than
once would lead to repeated instances of the format field.  Shortly
afterward, we noticed that it cached the Custom Fields that were added
in BuildFormatString and removed the %once block, breaking
SetFieldsOnce.  Worked in d77874ed broken in c07d62ad when fixing other
bugs.  Hasn't worked in a single released version of RT.